requirements.txt
streamlit
openai
langchain
langchain-core
langchain-text-splitters
langchain-community
langchain-openai
langchain-chains
faiss-cpu
pypdf2
python-docx
tiktoken
python-dotenv